requirements.txt
asgiref
boto3
botocore
certifi
chardet
charset-normalizer
defusedxml
diff-match-patch
Django
django-import-export
django-phonenumber-field
django-storages
et-xmlfile
gunicorn
idna
jmespath
MarkupPy
odfpy
openpyxl
packaging
phonenumbers
Pillow
pyotp
pypng
python-dateutil
pytz
PyYAML
qrcode
razorpay
reportlab
requests
s3transfer
segno
setuptools
six
sqlparse
tablib
typing_extensions
urllib3
xlrd
xlwt
